pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.171.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.6",
    "pymupdf>=1.26.0",
    "python-dateutil>=2.9.0.post0",
    "streamlit>=1.45.1",
]
